+++ changelog.md
@@ -1,6 +1,7 @@
 # changelog
 
 ## 1.18.0
+* `NEW` `LuaDoc`: supports `---@diagnostic disable`
 * `CHG` `Windows`: dose not provide `ucrt` any more
 * `CHG` `Lua.workspace.library`: use `path[]` instead of `<path, true>`
 * `FIX` missed syntax error `local a <const>= 1`